% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lauch_app.R
\name{launch_shiny_app}
\alias{launch_shiny_app}
\title{Application Launcher}
\usage{
launch_shiny_app(app = "shiny-snExplorer", use_browser = TRUE)
}
\arguments{
\item{app}{name of application.}

\item{use_browser}{logical.}
}
\description{
Application Launcher
}
\examples{
\dontrun{


launch_shiny_app()
}

}
